requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
Django==3.1
django-phonenumber-field==5.0.0
django-recaptcha3==0.4.0
idna==2.10
phonenumbers==8.12.20
Pillow==8.1.2
psycopg2-binary==2.8.6
python-decouple==3.4
pytils==0.3
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
